v3, ajout  EMD depuis POT PYthon

+++ emd.h
@@ -29,7 +29,7 @@ enum ProblemType {
 };
 
 
-int EMD_wrap(int n1,int n2, double *X, double *Y,double *D, double *G, double* alpha, double* beta, int maxIter);
+double EMD_wrap(int n1,int n2, double *X, double *Y,double *D, double *G, double* alpha, double* beta, int maxIter);
 
 int EMD_nb(int n1, int n2, double *X, double *Y, double *D, double *G,
            double* alpha, double* beta, int maxIter);
